Move comments array to PostDetailModal
